feat: add invoice services and actions; adjust navbar layout

+++ server/services/invoice.service.ts
@@ -2,61 +2,75 @@ import { unstable_cache } from "next/cache"
 import sql from "../db"
 
 interface InvoiceType {
-    id_invoice_type: number,
-    name: string,
-    created_at: Date
+    id_invoice_type?: number | null,
+    name?: string | null,
+    created_at?: Date | null
 }
 
 // ------------------- TYPES
 
 // FIND
 const findTypes = unstable_cache(
-    async (): Promise<InvoiceType[]> => await sql<InvoiceType[]>`
-        select * from ana_invoice_types`,
+    async () => {
+        const data = await sql`select * from ana_invoice_types`
+
+        return data.map((d) => ({
+            ...d,
+            id_invoice_type: Number(d.id_invoice_type)
+        }))
+    },
     ['invoices-findTypes'],
     { tags: ['invoices'] }
 )
 
 // EXISTY
-
-
-// CREATE
-const createType = unstable_cache(
+const extistsType = unstable_cache(
     async (name: string) => {
-
-        await sql`
-            insert into ana_invoice_types (name) values
-            (${name})
+        const [data] = await sql`
+            select * from ana_invoice_types
+            where lower(name) = lower(${name})
         `
+        return data
     },
-    ['invoices-createType'],
+    ['invoices-existsType'],
     { tags: ['invoices'] }
 )
 
-// UPDATE
-const updateType = unstable_cache(
-    async ({ id_invoice_type, name }: InvoiceType) => {
+// CREATE
+async function createType(name: string) {
+    await sql`
+        insert into ana_invoice_types (name) values
+        (${name})
+    `
+}
 
-        const [data] = await sql<InvoiceType[]>`
-            update ana_invoice_types set
-            name = ${name}
-            where id = ${id_invoice_type}
-        `
-    }
-)
+// UPDATE
+async function updateType(params: InvoiceType) {
+    const { id_invoice_type, name } = params
+
+    if (!id_invoice_type) throw new Error('id ausente')
+    if (!name) throw new Error('nome ausente')
+
+    await sql`
+        update ana_invoice_types set
+        name = ${name}
+        where id_invoice_type = ${id_invoice_type}
+    `
+}
 
 // DELETE
-const deleteType = unstable_cache(
-    async ({ id_invoice_type }: InvoiceType) => {
-        await sql`
-            delete from ana_invoice_types
-            where id_invoice_type = ${id_invoice_type}
-        `
-    }
-)
+async function deleteType({ id_invoice_type }: InvoiceType) {
+    if (!id_invoice_type) throw new Error('id ausente')
+
+    await sql`
+        delete from ana_invoice_types
+        where id_invoice_type = ${id_invoice_type}
+    `
+}
 
 const invoiceService = {
     findTypes,
+    extistsType,
     createType,
     updateType,
     deleteType
